fixed location of test_simulator
